move gkr `verify()` to air crate
