+++ air/Cargo.toml
@@ -31,6 +31,7 @@ internals = []
 
 [dependencies]
 thiserror = { version = "1.0", default-features = false }
+static_assertions = "1.1.0"
 vm-core = { package = "miden-core", path = "../core", version = "0.9", default-features = false }
 winter-air = { package = "winter-air", version = "0.9", default-features = false }
 winter-prover = { package = "winter-prover", version = "0.9", default-features = false }
