+++ air/src/gkr_proof.rs
@@ -2,11 +2,21 @@
 use core::ops::{Add, Index};
 
 use alloc::vec::Vec;
-use vm_core::{polynom, FieldElement};
+use static_assertions::const_assert;
+use vm_core::{polynom, Felt, FieldElement};
 use winter_air::{GkrRandElements, GkrVerifier};
 use winter_prover::{
     crypto::{ElementHasher, RandomCoin},
     Deserializable, Serializable,
+};
+
+use crate::{
+    decoder::{DECODER_OP_BITS_OFFSET, DECODER_USER_OP_HELPERS_OFFSET},
+    trace::{
+        chiplets::{MEMORY_D0_COL_IDX, MEMORY_D1_COL_IDX},
+        range::{M_COL_IDX, V_COL_IDX},
+    },
+    CHIPLETS_OFFSET, TRACE_WIDTH,
 };
 
 pub struct GkrCircuitVerifier {}
@@ -323,6 +333,74 @@ impl<E: FieldElement> Index<usize> for MultiLinearPoly<E> {
     }
 }
 
+/// The EQ (equality) function is the binary function defined by
+///
+/// EQ:    {0 , 1}^ν ⛌ {0 , 1}^ν ⇾ {0 , 1}
+///   ((x_0, ..., x_{ν - 1}), (y_0, ..., y_{ν - 1})) ↦ \prod_{i = 0}^{ν - 1} (x_i * y_i + (1 - x_i)
+/// * (1 - y_i))
+///
+/// Taking It's multi-linear extension EQ^{~}, we can define a basis for the set of multi-linear
+/// polynomials in ν variables by
+///         {EQ^{~}(., (y_0, ..., y_{ν - 1})): (y_0, ..., y_{ν - 1}) ∈ {0 , 1}^ν}
+/// where each basis function is a function of its first argument. This is called the Lagrange or
+/// evaluation basis with evaluation set {0 , 1}^ν.
+///
+/// Given a function (f: {0 , 1}^ν ⇾ 𝔽), its multi-linear extension (i.e., the unique
+/// mult-linear polynomial extending f to (f^{~}: 𝔽^ν ⇾ 𝔽) and agrees with it on {0 , 1}^ν) is
+/// defined as the summation of the evaluations of f against the Lagrange basis.
+/// More specifically, given (r_0, ..., r_{ν - 1}) ∈ 𝔽^ν, then:
+///
+///     f^{~}(r_0, ..., r_{ν - 1}) = \sum_{(y_0, ..., y_{ν - 1}) ∈ {0 , 1}^ν}
+///                  f(y_0, ..., y_{ν - 1}) EQ^{~}((r_0, ..., r_{ν - 1}), (y_0, ..., y_{ν - 1}))
+///
+/// We call the Lagrange kernel the evaluation of the EQ^{~} function at
+/// ((r_0, ..., r_{ν - 1}), (y_0, ..., y_{ν - 1})) for all (y_0, ..., y_{ν - 1}) ∈ {0 , 1}^ν for
+/// a fixed (r_0, ..., r_{ν - 1}) ∈ 𝔽^ν.
+///
+/// [`EqFunction`] represents EQ^{~} the mult-linear extension of
+///
+/// ((y_0, ..., y_{ν - 1}) ↦ EQ((r_0, ..., r_{ν - 1}), (y_0, ..., y_{ν - 1})))
+///
+/// and contains a method to generate the Lagrange kernel for defining evaluations of multi-linear
+/// extensions of arbitrary functions (f: {0 , 1}^ν ⇾ 𝔽) at a given point (r_0, ..., r_{ν - 1})
+/// as well as a method to evaluate EQ^{~}((r_0, ..., r_{ν - 1}), (t_0, ..., t_{ν - 1}))) for
+/// (t_0, ..., t_{ν - 1}) ∈ 𝔽^ν.
+pub struct EqFunction<E> {
+    r: Vec<E>,
+}
+
+impl<E: FieldElement> EqFunction<E> {
+    /// Creates a new [EqFunction].
+    pub fn new(r: Vec<E>) -> Self {
+        let tmp = r.clone();
+        EqFunction { r: tmp }
+    }
+
+    /// Computes EQ((r_0, ..., r_{ν - 1}), (t_0, ..., t_{ν - 1}))).
+    pub fn evaluate(&self, t: &[E]) -> E {
+        assert_eq!(self.r.len(), t.len());
+
+        (0..self.r.len())
+            .map(|i| self.r[i] * t[i] + (E::ONE - self.r[i]) * (E::ONE - t[i]))
+            .fold(E::ONE, |acc, term| acc * term)
+    }
+
+    /// Computes EQ((r_0, ..., r_{ν - 1}), (y_0, ..., y_{ν - 1})) for all
+    /// (y_0, ..., y_{ν - 1}) ∈ {0 , 1}^ν i.e., the Lagrange kernel at r = (r_0, ..., r_{ν - 1}).
+    pub fn evaluations(&self) -> Vec<E> {
+        compute_lagrange_basis_evals_at(&self.r)
+    }
+
+    /// Returns the evaluations of
+    /// ((y_0, ..., y_{ν - 1}) ↦ EQ^{~}((r_0, ..., r_{ν - 1}), (y_0, ..., y_{ν - 1})))
+    /// over {0 , 1}^ν.
+    pub fn ml_at(evaluation_point: Vec<E>) -> MultiLinearPoly<E> {
+        let eq_evals = EqFunction::new(evaluation_point.clone()).evaluations();
+        MultiLinearPoly::from_evaluations(eq_evals)
+            .expect("should not fail because evaluations is a power of two")
+    }
+}
+
 /// The coefficients of a univariate polynomial of degree n with the linear term coefficient
 /// omitted.
 #[derive(Clone, Debug)]
@@ -349,6 +427,195 @@ impl<E: FieldElement> UnivariatePolyCoef<E> {
         // evaluate
         polynom::eval(&complete_coefficients, *challenge)
     }
+}
+
+/// A multi-variate polynomial for composing individual multi-linear polynomials.
+pub trait CompositionPolynomial<E: FieldElement> {
+    /// Maximum degree in all variables.
+    fn max_degree(&self) -> u32;
+
+    /// Given a query, of length equal the number of variables, evaluates [Self] at this query.
+    fn evaluate(&self, query: &[E]) -> E;
+}
+
+/// A composition polynomial used in the GKR protocol for all of its sum-checks except the final
+/// one.
+#[derive(Clone)]
+pub struct GkrComposition<E>
+where
+    E: FieldElement<BaseField = Felt>,
+{
+    pub combining_randomness: E,
+}
+
+impl<E> GkrComposition<E>
+where
+    E: FieldElement<BaseField = Felt>,
+{
+    pub fn new(combining_randomness: E) -> Self {
+        Self {
+            combining_randomness,
+        }
+    }
+}
+
+impl<E> CompositionPolynomial<E> for GkrComposition<E>
+where
+    E: FieldElement<BaseField = Felt>,
+{
+    fn max_degree(&self) -> u32 {
+        3
+    }
+
+    fn evaluate(&self, query: &[E]) -> E {
+        let eval_left_numerator = query[0];
+        let eval_right_numerator = query[1];
+        let eval_left_denominator = query[2];
+        let eval_right_denominator = query[3];
+        let eq_eval = query[4];
+        eq_eval
+            * ((eval_left_numerator * eval_right_denominator
+                + eval_right_numerator * eval_left_denominator)
+                + eval_left_denominator * eval_right_denominator * self.combining_randomness)
+    }
+}
+
+/// A composition polynomial used in the GKR protocol for its final sum-check.
+#[derive(Clone)]
+pub struct GkrCompositionMerge<E>
+where
+    E: FieldElement<BaseField = Felt>,
+{
+    pub sum_check_combining_randomness: E,
+    pub tensored_merge_randomness: Vec<E>,
+    pub log_up_randomness: Vec<E>,
+}
+
+impl<E> GkrCompositionMerge<E>
+where
+    E: FieldElement<BaseField = Felt>,
+{
+    pub fn new(
+        combining_randomness: E,
+        merge_randomness: Vec<E>,
+        log_up_randomness: Vec<E>,
+    ) -> Self {
+        let tensored_merge_randomness =
+            EqFunction::ml_at(merge_randomness.clone()).evaluations().to_vec();
+
+        Self {
+            sum_check_combining_randomness: combining_randomness,
+            tensored_merge_randomness,
+            log_up_randomness,
+        }
+    }
+}
+
+impl<E> CompositionPolynomial<E> for GkrCompositionMerge<E>
+where
+    E: FieldElement<BaseField = Felt>,
+{
+    fn max_degree(&self) -> u32 {
+        // Computed as:
+        // 1 + max(left_numerator_degree + right_denom_degree, right_numerator_degree +
+        // left_denom_degree)
+        5
+    }
+
+    fn evaluate(&self, query: &[E]) -> E {
+        let [numerators, denominators] =
+            evaluate_fractions_at_main_trace_query(query, &self.log_up_randomness);
+
+        let numerators = MultiLinearPoly::from_evaluations(numerators.to_vec()).unwrap();
+        let denominators = MultiLinearPoly::from_evaluations(denominators.to_vec()).unwrap();
+
+        let (left_numerators, right_numerators) = numerators.project_least_significant_variable();
+        let (left_denominators, right_denominators) =
+            denominators.project_least_significant_variable();
+
+        let eval_left_numerators =
+            left_numerators.evaluate_with_lagrange_kernel(&self.tensored_merge_randomness);
+        let eval_right_numerators =
+            right_numerators.evaluate_with_lagrange_kernel(&self.tensored_merge_randomness);
+
+        let eval_left_denominators =
+            left_denominators.evaluate_with_lagrange_kernel(&self.tensored_merge_randomness);
+        let eval_right_denominators =
+            right_denominators.evaluate_with_lagrange_kernel(&self.tensored_merge_randomness);
+
+        let eq_eval = query[TRACE_WIDTH];
+
+        eq_eval
+            * ((eval_left_numerators * eval_right_denominators
+                + eval_right_numerators * eval_left_denominators)
+                + eval_left_denominators
+                    * eval_right_denominators
+                    * self.sum_check_combining_randomness)
+    }
+}
+
+/// Defines the number of wires in the input layer that are generated from a single main trace row.
+pub const NUM_WIRES_PER_TRACE_ROW: usize = 8;
+const_assert!(NUM_WIRES_PER_TRACE_ROW.is_power_of_two());
+
+/// Converts a main trace row (or more generally "query") to numerators and denominators of the
+/// input layer.
+pub fn evaluate_fractions_at_main_trace_query<E>(
+    query: &[E],
+    log_up_randomness: &[E],
+) -> [[E; NUM_WIRES_PER_TRACE_ROW]; 2]
+where
+    E: FieldElement,
+{
+    // numerators
+    let multiplicity = query[M_COL_IDX];
+    let f_m = {
+        let mem_selec0 = query[CHIPLETS_OFFSET];
+        let mem_selec1 = query[CHIPLETS_OFFSET + 1];
+        let mem_selec2 = query[CHIPLETS_OFFSET + 2];
+        mem_selec0 * mem_selec1 * (E::ONE - mem_selec2)
+    };
+
+    let f_rc = {
+        let op_bit_4 = query[DECODER_OP_BITS_OFFSET + 4];
+        let op_bit_5 = query[DECODER_OP_BITS_OFFSET + 5];
+        let op_bit_6 = query[DECODER_OP_BITS_OFFSET + 6];
+
+        (E::ONE - op_bit_4) * (E::ONE - op_bit_5) * op_bit_6
+    };
+
+    // denominators
+    let alphas = log_up_randomness;
+
+    let table_denom = alphas[0] - query[V_COL_IDX];
+    let memory_denom_0 = -(alphas[0] - query[MEMORY_D0_COL_IDX]);
+    let memory_denom_1 = -(alphas[0] - query[MEMORY_D1_COL_IDX]);
+    let stack_value_denom_0 = -(alphas[0] - query[DECODER_USER_OP_HELPERS_OFFSET]);
+    let stack_value_denom_1 = -(alphas[0] - query[DECODER_USER_OP_HELPERS_OFFSET + 1]);
+    let stack_value_denom_2 = -(alphas[0] - query[DECODER_USER_OP_HELPERS_OFFSET + 2]);
+    let stack_value_denom_3 = -(alphas[0] - query[DECODER_USER_OP_HELPERS_OFFSET + 3]);
+
+    [
+        [multiplicity, f_m, f_m, f_rc, f_rc, f_rc, f_rc, E::ZERO],
+        [
+            table_denom,
+            memory_denom_0,
+            memory_denom_1,
+            stack_value_denom_0,
+            stack_value_denom_1,
+            stack_value_denom_2,
+            stack_value_denom_3,
+            E::ONE,
+        ],
+    ]
+}
+
+/// Contains the round challenges sent by the Verifier up to some round as well as the current
+/// reduced claim.
+#[derive(Debug)]
+pub struct SumCheckRoundClaim<E: FieldElement> {
+    pub eval_point: Vec<E>,
+    pub claim: E,
 }
 
 // HELPER
