+++ processor/Cargo.toml
@@ -27,7 +27,6 @@ tracing = { version = "0.1", default-features = false, features = [
     "attributes",
 ] }
 miden-air = { package = "miden-air", path = "../air", version = "0.9", default-features = false }
-static_assertions = "1.1.0"
 thiserror = { version = "1.0", default-features = false }
 vm-core = { package = "miden-core", path = "../core", version = "0.9", default-features = false }
 winter-prover = { package = "winter-prover", version = "0.9", default-features = false }
