+++ processor/src/lib.rs
@@ -58,8 +58,7 @@ use chiplets::Chiplets;
 mod trace;
 use trace::TraceFragment;
 pub use trace::{
-    prove_virtual_bus, verify_virtual_bus, ChipletsLengths, ExecutionTrace, TraceLenSummary,
-    NUM_RAND_ROWS,
+    prove_virtual_bus, ChipletsLengths, ExecutionTrace, TraceLenSummary, NUM_RAND_ROWS,
 };
 
 mod errors;
