+++ processor/src/trace/mod.rs
@@ -21,7 +21,7 @@ mod utils;
 pub use utils::{AuxColumnBuilder, ChipletsLengths, TraceFragment, TraceLenSummary};
 
 mod virtual_bus;
-pub use virtual_bus::{prove as prove_virtual_bus, verify as verify_virtual_bus};
+pub use virtual_bus::prove as prove_virtual_bus;
 
 #[cfg(test)]
 mod tests;
