+++ processor/src/trace/virtual_bus/circuit/error.rs
@@ -1,5 +1,4 @@
 use crate::trace::virtual_bus::sum_check::SumCheckProverError;
-use crate::trace::virtual_bus::sum_check::SumCheckVerifierError;
 
 #[derive(Debug, thiserror::Error)]
 pub enum ProverError {
@@ -9,16 +8,4 @@ pub enum ProverError {
     FailedToProveSumCheck(#[from] SumCheckProverError),
     #[error("failed to generate the random challenge")]
     FailedToGenerateChallenge,
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum VerifierError {
-    #[error("one of the claimed circuit denominators is zero")]
-    ZeroOutputDenominator,
-    #[error("the output of the fraction circuit is not equal to the expected value")]
-    MismatchingCircuitOutput,
-    #[error("failed to generate the random challenge")]
-    FailedToGenerateChallenge,
-    #[error("failed to verify the sum-check proof")]
-    FailedToVerifySumCheck(#[from] SumCheckVerifierError),
 }
