--- processor/src/trace/virtual_bus/multilinear/error.rs
@@ -1,1 +1,0 @@
-
